with game manager and ballon spawns
please pay attention that the current implamantaion destroy all the ballons at the end of each wave. when we hit ballon with arrow the object just setActive(false) and not destroyed so we won't get null exceptions when destorying the ballons.
also ballon number is decided according to base factor of (currently 15) multiple by the wave number as same as the enemies